Revert "Refactor tests to use Half for image pixel data"
This reverts commit 2e440a274e750e2b5d1fa4cd660f60c9d1025994.

+++ PixelsorterTests/SorterTests/SorterTests.cs
@@ -28,7 +28,7 @@ namespace Pixelsorter.Tests.SorterTests
                 ..SorterTestHelpers.Gray, ..SorterTestHelpers.LowSaturation, ..SorterTestHelpers.MidSaturation, ..SorterTestHelpers.HighSaturation
             ]).reshape(4, 4, 3);
 
-            Assert.Equal(expectedData.ToArray<Half>(), sortedImage.ToArray<Half>());
+            Assert.Equal(expectedData.ToArray<float>(), sortedImage.ToArray<float>());
         }
 
         [Fact]
@@ -45,7 +45,7 @@ namespace Pixelsorter.Tests.SorterTests
                 ..SorterTestHelpers.HighSaturation, ..SorterTestHelpers.MidSaturation, ..SorterTestHelpers.LowSaturation, ..SorterTestHelpers.Gray
             ]).reshape(4, 4, 3);
 
-            Assert.Equal(expectedData.ToArray<Half>(), sortedImage.ToArray<Half>());
+            Assert.Equal(expectedData.ToArray<float>(), sortedImage.ToArray<float>());
         }
 
         [Fact]
@@ -62,7 +62,7 @@ namespace Pixelsorter.Tests.SorterTests
                 ..SorterTestHelpers.HighSaturation, ..SorterTestHelpers.HighSaturation, ..SorterTestHelpers.HighSaturation, ..SorterTestHelpers.HighSaturation
             ]).reshape(4, 4, 3);
 
-            Assert.Equal(expectedData.ToArray<Half>(), sortedImage.ToArray<Half>());
+            Assert.Equal(expectedData.ToArray<float>(), sortedImage.ToArray<float>());
         }
 
         [Fact]
@@ -79,7 +79,7 @@ namespace Pixelsorter.Tests.SorterTests
                 ..SorterTestHelpers.Gray, ..SorterTestHelpers.Gray, ..SorterTestHelpers.Gray, ..SorterTestHelpers.Gray
             ]).reshape(4, 4, 3);
 
-            Assert.Equal(expectedData.ToArray<Half>(), sortedImage.ToArray<Half>());
+            Assert.Equal(expectedData.ToArray<float>(), sortedImage.ToArray<float>());
         }
 
         [Fact]
@@ -96,7 +96,7 @@ namespace Pixelsorter.Tests.SorterTests
                 ..SorterTestHelpers.Gray, ..SorterTestHelpers.MidSaturation, ..SorterTestHelpers.HighSaturation, ..SorterTestHelpers.HighSaturation
             ]).reshape(4, 4, 3);
 
-            Assert.Equal(expectedData.ToArray<Half>(), sortedImage.ToArray<Half>());
+            Assert.Equal(expectedData.ToArray<float>(), sortedImage.ToArray<float>());
         }
 
         [Fact]
@@ -113,7 +113,7 @@ namespace Pixelsorter.Tests.SorterTests
                 ..SorterTestHelpers.MidSaturation, ..SorterTestHelpers.MidSaturation, ..SorterTestHelpers.HighSaturation, ..SorterTestHelpers.LowSaturation
             ]).reshape(4, 4, 3);
 
-            Assert.Equal(expectedData.ToArray<Half>(), sortedImage.ToArray<Half>());
+            Assert.Equal(expectedData.ToArray<float>(), sortedImage.ToArray<float>());
         }
 
         [Fact]
